requirements.txt
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Admin==1.6.1
email_validator==2.1.0.post1

# For split testing
Flask-Split==0.3.0

# For content editing
Flask-CKEditor==0.4.6
beautifulsoup4==4.12.2

# For analytics and data handling
numpy==1.26.0
pandas==2.1.1
matplotlib==3.8.0
scipy==1.11.3

# For security
Flask-Bcrypt==1.0.1
Flask-Talisman==1.1.0

# For API
Flask-RESTful==0.3.10

# For file handling
Werkzeug==2.3.7
python-dotenv==1.0.0

# For testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0